gateway ended

+++ src/main/java/ir/gambaloo/module/TaxGateway.java
@@ -3,7 +3,7 @@ package ir.gambaloo.module;
 public class TaxGateway extends Gateway{
     private static final double taxRate=0.05;
 
-    public TaxGateway(int price) {
+    public TaxGateway(double price) {
         super(price+price*taxRate);
     }
